move Flags and Double bit into measurement module
