+++ dnp3/examples/outstation_tcp_server.rs
@@ -1,6 +1,5 @@
 use dnp3::app::measurement::*;
 use dnp3::app::CommandStatus;
-use dnp3::app::Flags;
 use dnp3::decode::*;
 use dnp3::link::{EndpointAddress, LinkErrorMode};
 use dnp3::outstation::config::OutstationConfig;
