+++ dnp3/src/app/gen/conversion.rs
@@ -11,7 +11,6 @@
 //
 
 use crate::app::measurement::*;
-use crate::app::*;
 use crate::app::variations::*;
 
 impl From<Group2Var2> for Binary {
