+++ dnp3/src/app/parse/bit.rs
@@ -1,5 +1,5 @@
+use crate::app::measurement::DoubleBit;
 use crate::app::parse::range::Range;
-use crate::app::types::DoubleBit;
 use crate::util::bit::Bitfield;
 use crate::util::cursor::{ReadCursor, ReadError};
