+++ dnp3/src/master/session.rs
@@ -2,12 +2,12 @@ use crate::app::format::write;
 use crate::app::format::write::start_request;
 use crate::app::parse::parser::Response;
 use crate::app::Control;
-use crate::app::LinkStatusResult;
 use crate::app::Sequence;
 use crate::app::Timeout;
 use crate::decode::DecodeLevel;
 use crate::link::error::LinkError;
 use crate::link::EndpointAddress;
+use crate::link::LinkStatusResult;
 use crate::master::association::{AssociationMap, Next};
 use crate::master::error::TaskError;
 use crate::master::messages::{MasterMsg, Message};
