+++ dnp3/src/master/tasks/mod.rs
@@ -7,9 +7,8 @@ pub(crate) mod time;
 use crate::app::format::write::HeaderWriter;
 use crate::app::parse::parser::{HeaderCollection, Response};
 use crate::app::FunctionCode;
-use crate::app::LinkStatusResult;
 use crate::app::ResponseHeader;
-use crate::link::EndpointAddress;
+use crate::link::{EndpointAddress, LinkStatusResult};
 use crate::master::association::Association;
 use crate::master::error::TaskError;
 use crate::master::handle::Promise;
