+++ dnp3/src/outstation/database/details/event/buffer.rs
@@ -989,8 +989,6 @@ mod tests {
 
     use super::*;
     use crate::app::measurement::*;
-    use crate::app::DoubleBit;
-    use crate::app::Flags;
 
     fn insert_events(buffer: &mut EventBuffer) {
         buffer
