+++ dnp3/src/outstation/database/details/range/static_db.rs
@@ -1,6 +1,4 @@
 use crate::app::measurement::*;
-use crate::app::DoubleBit;
-use crate::app::Flags;
 use crate::outstation::database::config::*;
 use crate::outstation::database::details::event::buffer::Insertable;
 use crate::outstation::database::details::range::traits::StaticVariation;
