+++ dnp3/src/outstation/database/details/range/traits.rs
@@ -1,8 +1,6 @@
 use crate::app::measurement::*;
 use crate::app::parse::traits::{FixedSize, FixedSizeVariation};
 use crate::app::variations::*;
-use crate::app::DoubleBit;
-use crate::app::Flags;
 use crate::outstation::database::config::*;
 use crate::util::bit::bits::{BIT_6, BIT_7};
 use crate::util::cursor::{WriteCursor, WriteError};
