+++ dnp3/src/outstation/database/details/range/writer.rs
@@ -1,5 +1,5 @@
+use crate::app::measurement::DoubleBit;
 use crate::app::variations::Variation;
-use crate::app::DoubleBit;
 use crate::app::QualifierCode;
 use crate::outstation::database::details::range::traits::{
     FixedWriteFn, ToBit, ToDoubleBit, WriteInfo, WriteType,
@@ -286,8 +286,7 @@ fn write_header(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::app::measurement::{Binary, DoubleBitBinary, Time};
-    use crate::app::Flags;
+    use crate::app::measurement::{Binary, DoubleBitBinary, Flags, Time};
     use crate::outstation::database::config::*;
     use crate::outstation::database::details::range::traits::StaticVariation;
