+++ ffi/dnp3-ffi/src/association.rs
@@ -1,7 +1,7 @@
 use crate::command::Command;
 use crate::ffi;
 use crate::request::Request;
-use dnp3::app::LinkStatusResult;
+use dnp3::link::LinkStatusResult;
 use dnp3::master::association::PollHandle;
 use dnp3::master::error::{CommandError, CommandResponseError, TimeSyncError};
 use dnp3::master::handle::AssociationHandle;
