+++ ffi/dnp3-ffi/src/outstation/database.rs
@@ -1,19 +1,7 @@
 use crate::ffi;
-use dnp3::app::measurement::{
-    Analog, AnalogOutputStatus, Binary, BinaryOutputStatus, Counter, DoubleBitBinary,
-    FrozenCounter, OctetString, Time,
-};
-use dnp3::app::Flags;
-use dnp3::app::{DoubleBit, Timestamp};
-use dnp3::outstation::database::config::{
-    AnalogConfig, AnalogOutputStatusConfig, BinaryConfig, BinaryOutputStatusConfig, CounterConfig,
-    DoubleBitBinaryConfig, EventAnalogOutputStatusVariation, EventAnalogVariation,
-    EventBinaryOutputStatusVariation, EventBinaryVariation, EventCounterVariation,
-    EventDoubleBitBinaryVariation, EventFrozenCounterVariation, FrozenCounterConfig,
-    OctetStringConfig, StaticAnalogOutputStatusVariation, StaticAnalogVariation,
-    StaticBinaryOutputStatusVariation, StaticBinaryVariation, StaticCounterVariation,
-    StaticDoubleBitBinaryVariation, StaticFrozenCounterVariation,
-};
+use dnp3::app::measurement::*;
+use dnp3::app::Timestamp;
+use dnp3::outstation::database::config::*;
 use dnp3::outstation::database::{Add, EventClass, EventMode, Remove, Update, UpdateOptions};
 
 pub use dnp3::outstation::database::Database;
